+++ src/librustc/rustc.rc
@@ -86,6 +86,10 @@ mod middle {
         #[legacy_exports]
         mod meth;
         #[legacy_exports]
+        mod cabi;
+        #[legacy_exports]
+        mod cabi_x86_64;
+        #[legacy_exports]
         mod foreign;
         #[legacy_exports]
         mod reflect;
